fix typo

+++ qiskit_addon_utils/exp_vals/expectation_values.py
@@ -92,7 +92,7 @@ def executor_expectation_values(
         ValueError if `meas_basis_axis` is `None` but `len(basis_dict) != 1`.
         ValueError if the number of entries in `basis_dict` does not equal the length of `bool_array` along `meas_basis_axis`.
     """
-    if isinstance(gamma_factor, int):
+    if isinstance(gamma_factor, float):
         _gamma_factor: np.ndarray | float = gamma_factor
     ##### VALIDATE INPUTS:
     avg_axis = _validate_avg_axis(avg_axis, len(bool_array.shape))
